added utils/step and added renderOnCondition util
