+++ components/FormWizard/FormWizard.jsx
@@ -6,27 +6,7 @@ import { useBeforeunload } from 'react-beforeunload';
 
 import DynamicStep from 'components/DynamicStep/DynamicStep';
 import Breadcrumbs from 'components/Breadcrumbs/Breadcrumbs';
-import Summary from 'components/Steps/Summary';
-import Confirmation from 'components/Steps/Confirmation';
-
-export const createSteps = (formSteps) => [
-  ...formSteps,
-  { id: 'summary', title: 'Summary', component: Summary },
-  { id: 'confirmation', title: 'Confirmation', component: Confirmation },
-];
-export const getNextStepPath = (
-  currentStepIndex,
-  steps,
-  formPath,
-  formData
-) => {
-  const nextStep = steps
-    .slice(currentStepIndex + 1)
-    .find((step) =>
-      step.conditionalRender ? step.conditionalRender(formData) : true
-    );
-  return `${formPath}${nextStep.id}`;
-};
+import { createSteps, getNextStepPath } from 'utils/steps';
 
 const FormWizard = ({
   formPath,
